Avoid group-char regex problem.
Was using group-char in commify-commas to remove training instance
of the group char at end of string.  But if the group-char is an RE
metacharacter, this would cause a problem.  E.g., with "." as the
group-char, it would remove a digit from the num string not just a
literal ".".  Replaced with s-chop-suffix, which avoids the problem.