Fix failure to commify numbers starting with currency symbol or delimiter.
E.g., $56504069 would not commify.  In the course of debugging this, it turned
out that this was part of a more general problem with how I extracted numbers
from the buffer.  I was backing up to the nearest non-blank, but that caused
my to grab the whole number, $56504069, and that failed the regex test for a
decimal number.

So, now we back up to the nearest non-blank, non-currency, or non-opening
delimiter.  This may need amendment for other cases in the future, so I'll
keep an open mind.